requirements.txt
azure-core==1.25.1
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
configparser==5.2.0
cryptography==37.0.4
et-xmlfile==1.1.0
idna==3.3
isodate==0.6.1
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.3
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rethinkdb==2.4.9
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
